improve: 字幕の部分をSubtitleに切り出した。
さらに歌詞の部分はロジックが複雑なため、Lyricsとして切り出した。

また、使わなくなったファイルを削除した。
