+++ app/frontend/components/PracticeBody.vue
@@ -5,109 +5,27 @@
             :song="song"
             :videoId="videoId"
         ></Video>
-        <v-row align="center" id="subtitle" justify="center">
-            <v-col cols="auto">
-                <v-tooltip top>
-                    <template v-slot:activator="{on}">
-                        <v-btn depressed fab color="white" @click="see" v-on="on">
-                            <v-icon x-large color="maccha">mdi-{{seeButton}}</v-icon>
-                        </v-btn>
-                    </template>
-                    <span>字幕ON/OFF</span>
-                </v-tooltip>
-            </v-col>
-            <v-col cols="9" class="px-0">
-                <v-sheet v-show="this.isVisible" color="white rounded-pill px-5 py-2" min-height="56px" id="lyrics-wrapper">
-                    <h2 class="text-center black--text" 
-                        v-for="line in lyricsLines" :key="line.time" v-show="showLyric(line)"
-                    >
-                        <!-- v-forで回ってきた1行の歌詞を、どのようなHTMLタグで作成するか条件分岐する -->
-                        <!-- しかしその前に、この1行の歌詞はそもままでは条件分岐できない（どの言葉がコールの部分か区別がないため） -->
-                        <!-- ゆえに一回splitedLyricメソッドで加工し、返ってきたもので条件分岐する -->
-                        <span v-for="(word, index) in splitedLyric(line.lyric)" :key="index">
-                            <!-- 歌詞がコール歌詞だと検出したら、特殊なクラスを付与する -->
-                            <span v-if="matchCallLyrics(word)"
-                                :style="{backgroundColor:callBgc}"
-                            >
-                                {{word}}
-                            </span>
-                            <!-- そうでない歌詞はそのまま展示する -->
-                            <span v-else>{{word}}</span>
-                        </span>
-                    </h2>
-                </v-sheet>
-            </v-col>
-            <v-col cols="auto">
-                <v-speed-dial direction="top" open-on-click transition="slide-y-reverse-transition" v-model="isSpeedDialActive">
-                    <template v-slot:activator>
-                        <v-btn depressed fab color="white">
-                            <v-icon x-large color="black">mdi-{{optionButton}}</v-icon>
-                        </v-btn>
-                    </template>
-                    <v-btn depressed fab color="white" @click.stop="rewind">
-                        <v-tooltip top>
-                            <template v-slot:activator="{on}">
-                                <v-icon x-large color="maccha" v-on="on">mdi-redo</v-icon>
-                            </template>
-                            <span>字幕0.5秒遅らせる</span>
-                        </v-tooltip>
-                    </v-btn>
-                    <v-btn depressed fab color="white" @click.stop="forward">
-                        <v-tooltip top>
-                            <template v-slot:activator="{on}">
-                                <v-icon x-large color="maccha" v-on="on">mdi-undo</v-icon>
-                            </template>
-                            <span>字幕0.5秒早める</span>
-                        </v-tooltip>
-                    </v-btn>
-                    <v-btn depressed fab color="white" @click.stop="like($event)">
-                        <v-tooltip top>
-                            <template v-slot:activator="{on}">
-                                <v-icon x-large color="pink" v-on="on">mdi-{{likeButton}}</v-icon>
-                            </template>
-                            <span>お気に入り</span>
-                        </v-tooltip>
-                    </v-btn>
-                    <v-btn depressed fab color="white" @click.stop>
-                        <v-tooltip top>
-                            <template v-slot:activator="{on}">
-                                <v-sheet v-on="on" color="transparent">
-                                    <v-swatches id="color-picker" 
-                                        v-model="callBgc" :swatches="swatches"
-                                        popover-x="left" close-on-select shapes="circles"
-                                    ></v-swatches>
-                                </v-sheet>
-                            </template>
-                            <span>コールの背景色</span>
-                        </v-tooltip>
-                    </v-btn>
-                </v-speed-dial>
-            </v-col>
-        </v-row>
+        <Subtitle
+            :lyricsLines="lyricsLines"
+        ></Subtitle>
     </v-container>
 </template>
 
 <script>
-    import '../images/icomoon/style.css'
-    import {hearts} from '../src/effects/hearts'
-    import VSwatches from 'vue-swatches'
-    import Pubsub from 'pubsub-js'
     import Video from './Video.vue'
+    import Subtitle from './Subtitle.vue'
+
     export default {
         name: "PracticeBody",
         components: {
             Video,
-            VSwatches
+            Subtitle,
         },
         data() {
             return {
                 song: "Make you happy",
                 artist: "NiziU",
-                isLiked: false,
-                isVisible: true,
-                isSpeedDialActive: false,
                 videoId: "CN11U5t83Ro",
-                currentTime: 0,
                 lyricsLines: [
                     {
                         "time": 1.694,
@@ -346,116 +264,7 @@
                         "lyric": " You tell me"
                     }
                 ],
-                
-                
-                callBgc: "#ff94ce",
-                swatches:[
-                    "#ff94ce", "#ff9eff", "#c1c1ff", "#99ffff", "#b2ffd8", "#d8ffb2", "#ffffb2", "#ffe0c1"
-                ],
-                timeOffset: 0,
             }
-        },
-        computed: {
-            likeButton(){
-                return this.isLiked ? "heart" : "heart-outline";
-            },
-            seeButton(){
-                return this.isVisible ? "eye" : "eye-off";
-            },
-            optionButton(){
-                return this.isSpeedDialActive ? "chevron-up" : "dots-vertical"
-            },
-            
-        },
-        methods: {
-            like(event){
-                if (!this.isLiked) {
-                    hearts(event.target);
-                }
-                this.isLiked = !this.isLiked;
-            },
-            see(){
-                this.isVisible = !this.isVisible;
-            },
-            
-            showLyric(line){
-                const nextLineIndex = this.lyricsLines.indexOf(line) + 1
-                // もし次のライン（歌詞）がない場合（つまりnextLineIndexと配列の要素数と同じ）：
-                if (nextLineIndex === this.lyricsLines.length &&
-                    this.currentTime + this.timeOffset >= line.time) {
-                    return true
-                }
-                const nextLine = this.lyricsLines[nextLineIndex]
-                if (this.currentTime + this.timeOffset >= line.time && 
-                    this.currentTime + this.timeOffset <= nextLine.time) {
-                    return true
-                }
-            },
-            splitedLyric(lyric){
-                // コール箇所（中括弧）にマッチする
-                const callReg = /【.*?】/g;
-                // コール箇所があれば歌詞の分断処理に入る、そうでなければそのまま配列に入れてreturn
-                if (callReg.test(lyric)) {
-                    // コール箇所以外にマッチする
-                    const notCallReg = /[^【】]+(?=【[^【】]+】)|[^【】]+$/g
-                    
-                    // コール箇所の歌詞だけが入っている配列、
-                    const callLyrics = lyric.match(callReg)
-                    // そうでない歌詞だけが入っている配列
-                    const notCallLyrics = lyric.match(notCallReg)
-
-                    // 上記両方を入れる容器を用意する
-                    const indexLyrics = {}
-    
-                    // まずはコール歌詞を入れていく、キー名はコール歌詞が元の歌詞で初めて現れたインデックス
-                    for(let i=0;i<callLyrics.length;i++){
-                        const index = lyric.indexOf(callLyrics[i])
-                        indexLyrics[index] = callLyrics[i]
-                    }
-                    for(let b=0;b<notCallLyrics.length;b++){
-                        const index = lyric.indexOf(notCallLyrics[b])
-                        indexLyrics[index] = notCallLyrics[b]
-                    }
-    
-                    // 上記の処理で全ての歌詞を「コールの歌詞」「そうでない歌詞」に分割できた
-                    // しかしindexLyricsはオブジェクトで、次のステップでは配列が望ましいので、下記にように変換する
-                    const resultArray = Object.keys(indexLyrics).map(function(key){
-                        return indexLyrics[key]
-                    })
-    
-                    // 結果をreturnする
-                    return resultArray
-                } else {
-                    return [lyric]
-                }
-            },
-            matchCallLyrics(word){
-                const callReg = /【.*?】/g;
-                return callReg.test(word)
-            },
-            rewind(){
-                this.timeOffset -= 0.5
-            },
-            forward(){
-                this.timeOffset += 0.5
-            },
-        },
-        mounted() {
-            // 第一引数はsubscribeの名前で、実際は使わないので「_」でポジションをとってくれている。
-            Pubsub.subscribe("catchVideoCurrentTime", (_, time)=>{
-                this.currentTime = time;
-            })
-        },
-        beforeDestroy() {
-            PubSub.clearAllSubscriptions();
         },
     }
 </script>
-
-<style scoped>
-    #color-picker{
-        left: -1.5px;
-        top: 2px;
-        background-color: transparent;
-    }
-</style>
