--- app/frontend/images/icomoon/selection.json
@@ -1,1 +1,0 @@
-{"IcoMoonType":"selection","icons":[{"icon":{"paths":["M533.333 128c198.4 0 366.080 129.28 425.387 308.053l-101.12 33.28c-44.8-136.107-173.227-234.667-324.267-234.667-83.627 0-159.147 30.72-218.453 80.213l111.787 111.787h-298.667v-298.667l110.933 110.933c78.933-68.267 181.333-110.933 294.4-110.933zM697.984 512l1.792 426.667h-122.624l-0.811-341.333h-79.915l-0.427-85.333h263.339z"],"attrs":[{}],"isMulticolor":false,"isMulticolor2":false,"grid":0,"tags":["rewind-1"]},"attrs":[{}],"properties":{"order":4,"id":1,"name":"rewind-1","prevSize":32,"code":59648},"setIdx":0,"setId":2,"iconIdx":0},{"icon":{"paths":["M551.851 512l0.597 426.667h-104.747l0.725-341.333h-83.84l-0.384-85.333h185.728zM490.667 128c113.067 0 215.467 42.667 294.4 110.933l110.933-110.933v298.667h-298.667l111.787-111.787c-59.307-49.493-134.827-80.213-218.453-80.213-151.040 0-279.467 98.56-324.267 234.667l-101.12-33.28c59.307-178.773 226.987-308.053 425.387-308.053z"],"attrs":[{}],"isMulticolor":false,"isMulticolor2":false,"grid":0,"tags":["fast-forward-1"]},"attrs":[{}],"properties":{"order":5,"id":0,"name":"fast-forward-1","prevSize":32,"code":59649},"setIdx":0,"setId":2,"iconIdx":1}],"height":1024,"metadata":{"name":"icomoon"},"preferences":{"showGlyphs":true,"showQuickUse":true,"showQuickUse2":true,"showSVGs":true,"fontPref":{"prefix":"icon-","metadata":{"fontFamily":"icomoon"},"metrics":{"emSize":1024,"baseline":6.25,"whitespace":50},"embed":false},"imagePref":{"prefix":"icon-","png":true,"useClassSelector":true,"color":0,"bgColor":16777215,"classSelector":".icon"},"historySize":50,"showCodes":true,"gridSize":16}}
