--- app/frontend/images/icomoon/style.css
@@ -1,33 +1,0 @@
-@font-face {
-  font-family: 'icomoon';
-  src:  url('fonts/icomoon.eot?9ilmab');
-  src:  url('fonts/icomoon.eot?9ilmab#iefix') format('embedded-opentype'),
-    url('fonts/icomoon.ttf?9ilmab') format('truetype'),
-    url('fonts/icomoon.woff?9ilmab') format('woff'),
-    url('fonts/icomoon.svg?9ilmab#icomoon') format('svg');
-  font-weight: normal;
-  font-style: normal;
-  font-display: block;
-}
-
-[class^="icon-"], [class*=" icon-"] {
-  /* use !important to prevent issues with browser extensions that change fonts */
-  font-family: 'icomoon' !important;
-  speak: never;
-  font-style: normal;
-  font-weight: normal;
-  font-variant: normal;
-  text-transform: none;
-  line-height: 1;
-
-  /* Better Font Rendering =========== */
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-.icon-rewind-1:before {
-  content: "\e900";
-}
-.icon-fast-forward-1:before {
-  content: "\e901";
-}
